Update mdns_lite and fix warnings
Updates mdns_lite to latest version and fixes usage of records. Use the
record constructor instead of manually constructing the record.

Fix warning about using deprecated `Mix.Config` (`Config` was added in
Elixir 1.9 and this project is configured with 1.12 as the minimum)

Remove unused query and actually use the timeout passed to `discover/1`